Change to use synchronous way to return the imei as constant
- Also renamed the class which was still named according to the forked
react-native-device-info
